refactor: relocate http folder
